Got basic loading working
Right now it's only through hard coding, but I'm able to set a json file and then load it.

The sprites are not dynamic like they should be, but the particles seem to behave correctly.

Resources load perfectly.

I need to make a way to call these through UI next instead of calling load on PostStartup